refactor(core): extract pure plugin utilities to principles-core
